Move plugin/grpc to plugin/ocgrpc
Due to the conflict between the gRPC Go package and this
plugin package name, rename the package.
