+++ plugin/ocgrpc/example_test.go
@@ -12,19 +12,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package grpctrace_test
+package ocgrpc_test
 
 import (
 	"log"
 
-	"go.opencensus.io/plugin/grpc/grpctrace"
+	"go.opencensus.io/plugin/ocgrpc"
+	"go.opencensus.io/plugin/ocgrpc/grpcstats"
 	"google.golang.org/grpc"
 )
 
 func ExampleNewClientStatsHandler() {
-	// Set up a new client connection with the OpenCensus
-	// stats handler to enable tracing for the outgoing requests.
-	conn, err := grpc.Dial("address", grpc.WithStatsHandler(grpctrace.NewClientStatsHandler()))
+	// Subscribe to collect client request count.
+	if err := grpcstats.RPCClientRequestCountView.Subscribe(); err != nil {
+		log.Fatal(err)
+	}
+
+	// Set up a connection to the server with the OpenCensus
+	// stats handler to enable stats and tracing.
+	conn, err := grpc.Dial("address", grpc.WithStatsHandler(ocgrpc.NewClientStatsHandler()))
 	if err != nil {
 		log.Fatalf("did not connect: %v", err)
 	}
@@ -32,8 +38,13 @@ func ExampleNewClientStatsHandler() {
 }
 
 func ExampleNewServerStatsHandler() {
-	// Set up a new client connection with the OpenCensus
-	// stats handler to enable tracing for the incoming requests.
-	s := grpc.NewServer(grpc.StatsHandler(grpctrace.NewServerStatsHandler()))
+	// Subscribe to collect server request count.
+	if err := grpcstats.RPCServerRequestCountView.Subscribe(); err != nil {
+		log.Fatal(err)
+	}
+
+	// Set up a new server with the OpenCensus
+	// stats handler to enable stats and tracing.
+	s := grpc.NewServer(grpc.StatsHandler(ocgrpc.NewServerStatsHandler()))
 	_ = s // use s
 }
